add helper text to topic dropdowns
