+++ packages/commonwealth/client/scripts/views/components/NewThreadFormModern/NewThreadForm.scss
@@ -141,4 +141,8 @@
   .JoinCommunityBanner {
     margin-top: 8px;
   }
+
+  .help-text-container {
+    margin-left: auto;
+  }
 }
